requirements.txt
streamlit
google-genai
PyPDF2
plotly
pandas